Setup for a new data set

Generate the three files needed for a new data set:

A file to import the raw data as an R data set

A file to generate the documentation of the data set

A file to contain the examples for the data generated and opens the file interactively if needed.

Usage

data_setup(data, open = rlang::is_interactive())

Arguments

data

character. Name of the data set for which the files are created. Must be an existing .dat file in the data-raw folder.

open

Whether to open the file for interactive editing.