Interval between cars on the M1 motorway

Times that 41 successive vehicles travelling northwards along the M1 motorway in England passed a fixed point near Junction 13 in Bedfordshire on Saturday 23 March 1985. The times are recorded on the 24-hour clock to the nearest second.

Usage

interval

Format

A data frame with 48 rows and 3 columns:

car

Unique ID to define the cars

time

Time at which the car passed the junction. Encoded as a string in the 'HH:MM:SS' format

Source

Lewis, T. and the M345 Course Team (1986) M345 Statistical Methods, Unit 2: Basic Methods: Testing and Estimation, Milton Keynes: The Open University, 16.

Details

The data were collected by Professor Toby Lewis as a sample of event times in a point process which it may be appropriate to model as a Poisson process. They are also ideal to learn to handle time objects in R.