Piston-ring failures

Number of failures of piston-rings in each of three legs in each of four steam-driven compressors located in the same building. The compressors have identical design and are oriented in the same way.

Usage

piston

Format

A data frame with X rows and X columns:

comp_id

Compressor ID (1 to 4)

leg

Compressor leg (North, Center, South)

n_failures

Number of failures observed)

Source

Davies, O.L. and Goldsmith, P.L. (eds.) (1972) Statistical Methods in Research and Production, 4th Edition, Edinburgh: Oliverand Boyd, 324.

Details

Questions of interest are whether the probability of failure varies between compressors or between different legs, and whether the pattern of the location of failures is different for different compressors. Davies and Goldsmith investigate these questions using chi-squared tests; log-linear analysis might be an alternative.

Examples

tab <- xtabs(n_failures ~ comp_id + leg, data = piston)
tab
#>        leg
#> comp_id Center North South
#>       1     17    17    12
#>       2      9    11    13
#>       3      8    11    19
#>       4      7    14    28

chisq.test(tab)
#> 
#> 	Pearson's Chi-squared test
#> 
#> data:  tab
#> X-squared = 11.722, df = 6, p-value = 0.06846
#>