Mortality and water hardness

Annual mortality rate per 100 000 for males, averaged over the years 1958-1964, and the calcium concentration (in parts per million) in the drinking water supply for 61 large towns in England and Wales.

Usage

water

Format

A data frame with 61 rows and 5 columns:

town

Name of the town, as a factor variable

mortality

Mortality rate per 100,000 inhabitants

calcium

Calcium concentration in the drinking water in parts per million (ppm)

lon

Longitude using the EPSG:4326 coordinate system

lat

Latitude using the EPSG:4326 coordinate system

Source

Hills, M. and the M345 Course Team (1986) M345 Statistical Methods, Unit 3: Examining Straight-line Data, Milton Keynes: The Open University, 28. Data provided by Professor M.J. Gardner, Medical Research Council Environmental Epiderniology Research Unit, Southampton.