Tensile strength of cement

Tensile strength of different batches of cement after different curing times. The tensile strength of cement depends on (among other things) the length of time for which the cement is dried or 'cured'. The relationship between curing time and strength is non-linear. Hald regresses log tensile strength on the reciprocal of curing time.

Usage

cement

Format

A data frame with 21 rows and 2 columns:

time

Curing time in days

strength

Tensile strength (kg/cm2)

Source

Hald, A. (1952) Statistical Theory with Engineering Applications, New York: John Wiley & Sons, 451.

Examples

library(ggplot2)
#> Error in library(ggplot2): there is no package called ‘ggplot2’
library(hsds)

# There is a clear non-linear relation between time and strength
# Hald regresses log tensile strength on the reciprocal of curing time.

ggplot(cement, aes(y = strength, x = time)) +
  geom_point() +
  geom_smooth(method = "lm", formula = y ~ exp(1 / x), se = FALSE) +
  theme_classic(base_size = 12)
#> Error in ggplot(cement, aes(y = strength, x = time)): could not find function "ggplot"

# Using Hald's method the relationship is clearly linear
ggplot(cement, aes(y = log(strength), x = 1 / time)) +
  geom_point() +
  geom_smooth(method = "lm", formula = y ~ x, se = FALSE) +
  theme_classic(base_size = 12)
#> Error in ggplot(cement, aes(y = log(strength), x = 1/time)): could not find function "ggplot"

# A linear model indeed fits well (high R-squared) and
# the reciprocal of time is significant
lm(log(strength) ~ I(1 / time), data = cement) |>
  summary()
#> 
#> Call:
#> lm(formula = log(strength) ~ I(1/time), data = cement)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -0.12376 -0.04601  0.01679  0.05098  0.10495 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  3.68782    0.02425  152.06  < 2e-16 ***
#> I(1/time)   -1.14553    0.05290  -21.66 7.47e-15 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 0.07557 on 19 degrees of freedom
#> Multiple R-squared:  0.9611,	Adjusted R-squared:  0.959 
#> F-statistic: 468.9 on 1 and 19 DF,  p-value: 7.473e-15
#>